requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dataiku-api-client==5.1.1
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.21.0
six==1.13.0
urllib3==1.24.3
Werkzeug==0.16.0
